# Feature/issue 396 app performance improvements