--- a/src/components/empty_component/empty_list_component.tsx
+++ b/src/components/empty_component/empty_list_component.tsx
@@ -1,31 +1,22 @@
 import React from 'react';
-import { StyleSheet } from 'react-native';
 import { Text, View } from 'native-base';
-import { colors, values } from '../../application/styles';
+import { colors } from '../../application/styles';
 
 export interface EmptyListProps {
     readonly message: JSX.Element;
 }
 
 export const EmptyListComponent: React.StatelessComponent<EmptyListProps> = (props: EmptyListProps): JSX.Element => (
-    <View style={{ flex: 1 }}>
-        <View style={styles.wrapper}>
-            <Text style={styles.message}>
+    <View
+        style={{
+            flex: 1,
+            paddingHorizontal: 10,
+            backgroundColor: colors.white,
+            alignItems: 'center',
+        }}
+    >
+            <Text style={{ color: colors.darkerGrey }}>
                 {props.message}
             </Text>
-        </View>
     </View>
 );
-
-const styles = StyleSheet.create({
-    wrapper: {
-        flexDirection: 'row',
-        backgroundColor: colors.white,
-        justifyContent: 'center',
-        borderRadius: values.lessRoundedBorderRadius,
-        padding: 5,
-    },
-    message: {
-        color: colors.darkerGrey,
-    },
-});