--- a/src/components/explore/explore_detail_component.tsx
+++ b/src/components/explore/explore_detail_component.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import { Content } from 'native-base';
+import { Trans } from '@lingui/react';
+import { View, Text } from 'native-base';
 import { ExploreSection } from '../../selectors/explore/types';
 import { RouterProps } from '../../application/routing';
 import { Id as TaskId, AddToSavedListAction, RemoveFromSavedListAction } from '../../stores/tasks';
-import { ExploreDetailHeadingComponent } from './explore_detail_heading_component';
 import { ExploreDetailContentComponent } from './explore_detail_content_component';
-import { ExploreDetailRelatedTasksComponent } from './explore_detail_related_tasks_component';
 import { TaskListItem } from '../../selectors/tasks/task_list_item';
-import { applicationStyles } from '../../application/styles';
+import { textStyles } from '../../application/styles';
+import { TaskListComponent, NoTasksAddedComponent } from '../tasks/task_list_component';
 
 export interface ExploreDetailProps {
     readonly section: ExploreSection;
@@ -25,19 +25,26 @@ type Props = ExploreDetailProps & ExploreDetailActions & RouterProps;
 export const ExploreDetailComponent: React.StatelessComponent<Props> =
     (props: Props): JSX.Element => {
         return (
-            <Content padder style={applicationStyles.body}>
-                <ExploreDetailHeadingComponent />
-                <ExploreDetailContentComponent
-                    section={props.section}
-                    sectionHasTasks={props.tasks.length > 0}
-                />
-                <ExploreDetailRelatedTasksComponent
-                    relatedTasks={props.tasks}
-                    history={props.history}
-                    savedTasksIdList={props.savedTasksIdList}
-                    addToSavedList={props.addToSavedList}
-                    removeFromSavedList={props.removeFromSavedList}
-                />
-            </Content>
+            <TaskListComponent
+                tasks={props.tasks}
+                savedTasksIdList={props.savedTasksIdList}
+                addToSavedList={props.addToSavedList}
+                removeFromSavedList={props.removeFromSavedList}
+                history={props.history}
+                emptyTaskListContent={<NoTasksAddedComponent />}
+                headerContent={<TaskListHeaderComponent {...props} />}
+            />
         );
     };
+
+const TaskListHeaderComponent = (props: Props): JSX.Element => (
+    <View>
+        <ExploreDetailContentComponent
+            section={props.section}
+            sectionHasTasks={props.tasks.length > 0}
+        />
+        <Text style={[textStyles.headlineH5StyleBlackLeft, { marginBottom: 3 }]}>
+            <Trans>EXPLORE TOPICS</Trans>
+        </Text>
+    </View>
+);