--- a/src/components/explore/explore_detail_component.tsx
+++ b/src/components/explore/explore_detail_component.tsx
@@ -38,7 +38,7 @@ export const ExploreDetailComponent: React.StatelessComponent<Props> =
     };
 
 const TaskListHeaderComponent = (props: Props): JSX.Element => (
-    <View>
+    <View padder>
         <ExploreDetailContentComponent
             section={props.section}
             sectionHasTasks={props.tasks.length > 0}