--- a/src/components/explore/explore_detail_related_tasks_component.tsx
+++ b/src/components/explore/explore_detail_related_tasks_component.tsx
@@ -32,6 +32,7 @@ export const ExploreDetailRelatedTasksComponent: React.StatelessComponent<Props>
                 addToSavedList={props.addToSavedList}
                 removeFromSavedList={props.removeFromSavedList}
                 history={props.history}
+                loadOnDemand={false}
             />
         </View>
     );