… src/components/explore/explore_detail_heading_component.tsx
… src/components/questionnaire/answer_component.tsx
… src/components/questionnaire/question_component.tsx
… src/components/questionnaire/questionnaire_component.tsx
… src/components/recommended_topics/recommended_topics_component.tsx
… src/components/tasks/task_detail_component.tsx
… src/components/tasks/task_detail_content_component.tsx
… src/components/tasks/task_detail_heading_component.tsx
… src/components/tasks/task_list_component.tsx
… src/components/tasks/task_list_item_component.tsx
… src/components/tasks_added/tasks_added_component.ts
… src/components/tasks_added/tasks_added_connected_component.ts
… src/selectors/questionnaire/get_show_questionnaire_popup.ts
… src/selectors/questionnaire/is_questionnaire_popup_shown.ts
… src/selectors/tasks/compute_newly_recommended_unsaved_tasks.ts
… src/selectors/tasks/select_newly_recommended_unsaved_tasks.ts