@bensinjin 2019-01-30
@rasmus-storjohann-PG  This PR grew, I did however need to find a solution to: https://github.com/pg-irc/pathways-frontend/issues/424, which I've done in my later commits (I can explain if need b), in order to properly move ahead with the styling stories coming.